chore: migrate remaining jobs into new format
