+++ web/package.json
@@ -18,11 +18,7 @@
     "test:watch": "dotenv -e ../.env -- jest --watch --runInBand",
     "test:e2e": "dotenv -e ../.env -- playwright test",
     "test:e2e:server": "dotenv -e ../.env -- jest --runInBand --detectOpenHandles --verbose --testPathPattern='__e2e__'",
-    "models:migrate": "dotenv -e ../.env -- tsx scripts/model-match.ts",
-    "generationCost:backfill": "dotenv -e ../.env -- tsx scripts/observations-backfill-calculated-cost.ts",
-    "clickhouseMigration:traces": "dotenv -e ../.env -- tsx scripts/migration-traces-postgres-to-clickhouse.ts",
-    "clickhouseMigration:observations": "dotenv -e ../.env -- tsx scripts/migration-observations-postgres-to-clickhouse.ts",
-    "clickhouseMigration:scores": "dotenv -e ../.env -- tsx scripts/migration-scores-postgres-to-clickhouse.ts"
+    "models:migrate": "dotenv -e ../.env -- tsx scripts/model-match.ts"
   },
   "dependencies": {
     "@anthropic-ai/tokenizer": "^0.0.4",
