+++ worker/src/backgroundMigrations/IBackgroundMigration.ts
@@ -1,7 +1,7 @@
 export interface IBackgroundMigration {
   validate: (
     args: Record<string, unknown>,
-  ) => Promise<{ valid: true; invalidReason: string | undefined }>;
+  ) => Promise<{ valid: boolean; invalidReason: string | undefined }>;
   run: (args: Record<string, unknown>) => Promise<void>;
   abort: () => Promise<void>;
 }
